Fix listening to presence channel broadcast
Hi,

This code wasn't working because Livewire try to call the `Echo.join(channel)[here/joining/leaving]()` but we can also listen to PresenceChannel with `.listen()`.
```
    protected $listeners = ['echo-presence:orders,OrderShipped' => 'notifyNewOrder'];
```

I fixed it by looking if the method exists on the `PresenceChannel` JS object (here, joining, leaving) but it may be preferable to do a clearer check with `if (event_name === 'here' && event_name === 'joining'…)`. I can modify the pull request if you want.